requirements.txt
openai
python-dotenv
smolagents 